AI-News-Summariser-main/requirements.txt
flask
flask-login
flask-sqlalchemy
pytesseract
validators
trafilatura
newspaper3k
transformers
deep-translator
requests
langdetect
torch
numpy
Pillow
python-dotenv
Werkzeug
nltk
beautifulsoup4
lxml
#sentencepiece
protobuf
regex
tqdm
packaging
filelock
typing-extensions
certifi
charset-normalizer
idna
urllib3
PyPDF2
python-magic-bin